First version

+++ src/main/java/org/ademun/timetableapi/repository/DayRepository.java
@@ -1,9 +1,10 @@
 package org.ademun.timetableapi.repository;
 
-import org.ademun.timetableapi.entity.Day;
+import org.ademun.timetableapi.model.Day;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DayRepository extends JpaRepository<Day, Integer> {
+public interface DayRepository extends JpaRepository<Day, Long> {
+
 }
