+++ src/main/java/org/ademun/timetableapi/repository/DisciplineRepository.java
@@ -1,9 +1,9 @@
 package org.ademun.timetableapi.repository;
 
-import org.ademun.timetableapi.entity.Discipline;
+import org.ademun.timetableapi.model.Discipline;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DisciplineRepository extends JpaRepository<Discipline, Integer> {
+public interface DisciplineRepository extends JpaRepository<Discipline, Long> {
 }
