+++ src/main/java/org/ademun/timetableapi/repository/OrderRepository.java
@@ -1,9 +1,9 @@
 package org.ademun.timetableapi.repository;
 
-import org.ademun.timetableapi.entity.Order;
+import org.ademun.timetableapi.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
